refactor(npc): split arrival_reactions.rs by concern (TD-010)
Behavior-preserving split of parish-npc arrival_reactions.rs (~1240 LOC)
into four submodules under arrival_reactions/:
- types.rs  — ReactionKind, NpcReaction, ArrivalContext
- templates.rs — template structs, defaults, placeholder substitution
- selection.rs — threshold, kind selection, capping, generation loop
- prompt.rs — LLM prompt construction and resolve_llm_greeting

Hub arrival_reactions.rs re-exports the unchanged public API.
Tests moved to arrival_reactions/tests.rs. architecture_fitness passes.
Part of #1200 (TD-010).
